Refactor file structure and move WMI wrapper to utils
This commit restructures the file paths for better organization, improves readability, and enhances the logical flow. The WMI wrapper previously situated in the root directory has been relocated to the utils directory to keep utility files together. In addition, this change also isolates platform-specific files within their corresponding platform directories to ensure better code maintainability. Extensive changes were also made in include files and the source CMakeLists.txt file was removed.

The WMI wrapper file has been refactored as it was unnecessarily cluttered. Its functions were moved to the `windows/utils/wmi_wrapper.h` file. This change promotes code cleanliness and contributes towards better maintenance in the future. All paths in the include statements were corrected to comply with the new structure.

Some changes were made to several include files - 'pragma once' directive was added where missing, to prevent multiple inclusions of the same header file. In addition, relative paths were adjusted to reflect the new structure.

Finally, the 'src/CMakeLists.txt' was removed because it had been already commented out and was not needed in this project.

The Example project builds on Windows.
