+++ include/hwinfo/hwinfo.h
@@ -3,10 +3,10 @@
 
 #pragma once
 
-#include <hwinfo/battery.h>
-#include <hwinfo/cpu.h>
-#include <hwinfo/disk.h>
-#include <hwinfo/gpu.h>
-#include <hwinfo/mainboard.h>
-#include <hwinfo/os.h>
-#include <hwinfo/ram.h>
+#include "battery.h"
+#include "cpu.h"
+#include "disk.h"
+#include "gpu.h"
+#include "mainboard.h"
+#include "os.h"
+#include "ram.h"
