+++ include/hwinfo/windows/ram.h
@@ -1,7 +1,9 @@
 // Copyright (c) Leon Freist <freist@informatik.uni-freiburg.de>
 // This software is part of HWBenchmark
 
-#include "hwinfo/platform.h"
+#pragma once
+
+#include "../platform.h"
 
 #ifdef HWINFO_WINDOWS
 #include <windows.h>
@@ -9,9 +11,8 @@
 #include <string>
 #include <vector>
 
-#include "../WMIwrapper.h"
-//#include "hwinfo/ram.h"
 #include "../utils/stringutils.h"
+#include "utils/wmi_wrapper.h"
 
 namespace hwinfo {
 namespace ram {
